Updated Time Class + Experiment Parameters
Fixed issue where runtime stats appeared as all zeroes by updating end_time only if start() is called.
Current time is used as endpoint in elapsed(), which calculates runtime even if stop() hasn't been called.
Increased the number of nodes to 1000, number of edges to 20000, and number of trials to 5000 to better see the differences between the {uniform, direct, inverse} strategies.